backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio>=0.21.0
instructor>=1.0.0
google-generativeai>=0.3.0
openai>=1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
email-validator==2.1.0
rapidfuzz==3.5.2

